Move QueryManipulator to AbstractRetriever's module
